Merged PR 52280: [41M7RQ] fix: avoid floating point issue after rotation
fix: avoid floating point issue after rotation
fix: Resolved the swiching unit issue, allowing delete all number in input field

https://link.orangelogic.com/Tasks/41M7RQ

Cause: When rotating by 90, 180, 270, or 360 degrees, the width and height of the asset should be swapped. However, we are applying the normal formula instead, which results in a floating-point precision issue.

Solution: If we are rotating by 90, 280, 270, or 360 degrees, swap the asset width and height.

https://link.orangelogic.com/Tasks/41M872

https://link.orangelogic.com/Tasks/41M844